✨ Added a custom hook for zustand store and useEffect
